+++ src/main/java/JUnitTest/TestJunit.java
@@ -4,7 +4,7 @@
  *   used to test if Junit works
  */
 
-package JUnitTests;
+package JUnitTest;
 
 import org.junit.Test;
 import static org.junit.Assert.assertEquals;
